requirements.txt
# Direct Python dependencies inferred from the repository source files.
edge-tts==7.2.8
google-genai==1.74.0
ollama
openai-whisper
pandas==3.0.2
pydantic==2.13.3
PyAudio
pyttsx3==2.99
reachy_mini==1.7.1
SpeechRecognition==3.16.1
